feat: rework resolver inside gnopls
