+++ pkg/resolver/main.go
@@ -1,0 +1,3 @@
+package resolver
+
+// package main
